rpi_uart_client/requirements.txt
# pip3 install -r requirements.txt
pyserial==3.5
